PostEntrega

+++ ft_cronos/ft_time_diff.c
@@ -1,6 +1,18 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   ft_time_diff.c                                     :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: uliherre <uliherre@student.42.fr>          +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2022/06/16 19:52:05 by uliherre          #+#    #+#             */
+/*   Updated: 2022/06/16 19:52:32 by uliherre         ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
 #include "../libft.h"
 
-long long	ft_time_diff(long long start, long long end)
+long int	ft_time_diff(long int start, long int end)
 {
 	return (end - start);
 }
